Created predict_unified_efield, which takes a given time and a trained pytorch NN and predicts the electric field and the electric potential, and creates some plots of the resulting values.
Updated unified_efield, such that it actually works. It now requires a file from sample_data that has downloaded the electric field, Kp, and Dst

Enhanced bin_5min such that it now takes any amount of time and creates 5 minute bins as long as the inputted time range is a multiple of 5 minutes. This is not a problem if any of the get_xxx_functions is used, as they automatically add whatever time is needed onto the end of the time range such that the time range is a multiple of 5 minutes

Fixed read_txt_files so that it works for all ranges of Dst time ranges

Updated Comments