Refactored linear_regression.py to become Neural_Networks.py, which is a storage file with all different architectures of neural networks.
Added the number of layers argument to unified_efield.py, so that the user can choose how many layers the NN has. Right now they cannot choose how many nodes are in each layer, but they can randomize the number. In the future I may implement choosing # of nodes per layer

Added the random argument in unified_efield.py, that allows the user to randomize the number of nodes in each layer of the resulting NN. Unified_efield also outputs parameters of the model, along with the test errors of the model, to a file called test_errors.txt

Changed remove_corot_efield to hopefully work for real this time

Updated Comments