Added additional optional arguments to create_neural_network.py, allowing the user to change the batch size, learning rate, and number of epochs.
Changed get_NN_inputs to actually get the last 5 minutes of data, even if the given data file is not continuous, and if the time cadence isn't 5 minutes (though the time cadence must have the 5 minute times in them. 5, 10 seconds are allowable, as are 1 and 2.5 minutes. However 13 seconds is not)

Edited create_histogram in some_plots, such that the cumulative histogram has the x and y axes switched. Known issues: The axes labels on the bar plot do not correspond with each bin